// web/src/api/gen/DiffLine.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiffLineKind } from "./DiffLineKind";

export type DiffLine = { type: DiffLineKind, 
/**
 * The line without its `+`, `-`, or space prefix and without the newline.
 */
text: string, old_no: number | null, new_no: number | null, };